+++ SimpleMDMpy/Logs.py
@@ -10,9 +10,24 @@ class Logs(SimpleMDMpy.SimpleMDM.Connection):
     def __init__(self, api_key):
         SimpleMDMpy.SimpleMDM.Connection.__init__(self, api_key)
         self.url = self._url("/logs")
-
-    def get_logs(self, id_override=0):
-        """And I mean all the LOGS"""
+    
+    def get_logs(self, starting_after=None, limit=None):
+        """Returns logs, and I mean all the LOGS
+        
+        Args:
+            starting_after (str, optional):  set to the id of the log object you
+                want to start with. Defaults to the first object.
+            limit (str, optional): A limit on the number of objects that will be
+                returned per API call. Setting this will still return all logs.
+                Defaults to 100.
+        
+        Returns:
+            array: An array of dictionary log objects.
+        """
         url = self.url
-        data = {}
-        return self._get_data(url, data, id_override=id_override)
+        params = {}
+        if starting_after:
+            params['starting_after'] = starting_after
+        if limit:
+            params['limit'] = limit
+        return self._get_data(url, params=params)