+++ tests/settings-sample.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+
+# add a SimpleMDM API key to use for tests
+api_key = ''
+
+# add the ID of a profile in your instance to test profile functions
+profile_id = '' # note what profile this is
+
+# add the id of a device in your instance to test functions against
+device_id = '' # note what device this is