+++ CHANGELOG.md
@@ -1,6 +1,42 @@
 # CHANGELOG
 
-## v3.0.6
+## [Unreleased]
+
+### Added
+- setup.cfg - Python package setup file ([#29](https://github.com/macadmins/simpleMDMpy/issues/29)) - TY [@bryanheinz](https://github.com/bryanheinz)
+- pyproject.toml - Python package meta setup file ([#29](https://github.com/macadmins/simpleMDMpy/issues/29)) - TY [@bryanheinz](https://github.com/bryanheinz)
+- tests - Added a few basic tests and including a readme on how to setup testing - TY [@bryanheinz](https://github.com/bryanheinz)
+
+### Changes
+
+- Added ability to update the actual device name via SimpleMDM ([#24](https://github.com/macadmins/simpleMDMpy/issues/24), [#38](https://github.com/macadmins/simpleMDMpy/issues/38)) - TY [@bryanheinz](https://github.com/bryanheinz)
+- Replaced get_logs() `id_override` input parameter with `starting_after` and `limit` ([#25](https://github.com/macadmins/simpleMDMpy/issues/25)) - TY [@bryanheinz](https://github.com/bryanheinz)
+- Fixes calls that return a single item ([#26](https://github.com/macadmins/simpleMDMpy/issues/26)) - TY [@MagerValp](https://github.com/MagerValp)
+- Add method to download profiles ([#40](https://github.com/macadmins/simpleMDMpy/issues/40)) - TY [@joncrain](https://github.com/joncrain)
+- Adds option for get_devices to include_awaiting_enrollment ([#43](https://github.com/macadmins/simpleMDMpy/issues/43)) - TY [@joncrain](https://github.com/joncrain)
+- Fixes `Devices.delete_device()` - TY [@MagerValp](https://github.com/MagerValp)
+- Add Devices methods for enabling/disabling remote desktop, and profile and user listing ([@MagerValp](https://github.com/MagerValp))
+- Add /devices request rate limiting to `_get_data` - TY [@MagerValp](https://github.com/MagerValp)
+- Add retry on 5xx errors to GET requests to `_get_data` - TY [@MagerValp](https://github.com/MagerValp)
+- Fixes `_get_data` so that it properly preserves all input parameters ([#45](https://github.com/macadmins/simpleMDMpy/issues/45)) - TY [@bryanheinz](https://github.com/bryanheinz)
+- Adds help docs to Devices.get_device() - TY [@bryanheinz](https://github.com/bryanheinz)
+- Add Scripts and ScriptJobs - TY [@MagerValp](https://github.com/MagerValp)
+- Fix pagination - TY [@jcfrt](https://github.com/jcfrt)
+- Fix rate limiting - TY [@MagerValp](https://github.com/MagerValp)
+
+### Issues
+
+- Closes issue #24
+- Closes issue #38
+- Closes issue #25
+- Closes issue #26
+- Closes issue #40
+- Closes issue #43
+- Closes issue #29
+- Closes issue #45
+- Closes issue #57
+
+## [v3.0.6]
 
 ### PRs Included
 
@@ -12,7 +48,7 @@
 
 - Add method to get all custom attributes for a device
 
-## v3.0.5
+## [v3.0.5]
 
 ### Issues
 
@@ -22,7 +58,7 @@
 
 - CODEOWNERS
 
-## v3.0.4
+## [v3.0.4]
 
 ### Issues
 
@@ -39,7 +75,7 @@
 - Merged with @MagerValp / simpleMDMpy @ [508540928](https://github.com/MagerValp/simpleMDMpy/commit/50854094bee2ac5306eded7c5614d76f3eab4c25)
 - minor tweaks on the readme
 
-## v3.0.3
+## [v3.0.3]
 
 ### Issues
 
@@ -57,7 +93,7 @@
 - default branch is now `main`
 - remove `data` payload from Devices.delete_device
 
-## v3.0.2
+## [v3.0.2]
 
 ### Issues
 
@@ -73,7 +109,7 @@
 
 - Changed paginaition to work without compounding to a `414`
 
-## v3.0.1
+## [v3.0.1]
 
 ### Issues
 
@@ -84,7 +120,7 @@
 - Changed paginaition to work, now returns obj not response
 - good catch @bryanheinz
 
-## v3.0.0
+## [v3.0.0]
 
 - Closes #3
 
@@ -93,7 +129,7 @@
 - removed forced encoding for `GET` responses
 - added some pylint comments
 
-## v2.1.0
+## [v2.1.0]
 
 ### Issues
 
@@ -103,7 +139,7 @@
 
 - fixed module names
 
-## v2.0.0
+## [v2.0.0]
 
 ### Issues